src/requirements.txt
requests
azure-data-tables
openai